fix rates table issue

+++ database/migrations/2022_02_26_230550_create_items_rates_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->tinyInteger('rate')->unsigned();
             $table->bigInteger('item_id')->unsigned();
             $table->foreign('item_id')->references('id')->on('items')->onDelete('cascade');
-            $table->integer('user_id')->unsigned()->nullable();
+            $table->bigInteger('user_id')->unsigned()->nullable();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
             $table->timestamp('created_at');
         });
